Organizational Work, ndarray SMU voltages
Renamed documentation folder to indicate that it was specifically for GPIB setup. Added folder for unit tests.

Enabled np.linspace() voltage input by casting previously-supported list-type voltage input as ndarray. This allows for checking for voltages == None in the measurement functions by using the numpy function voltages.all() == None, avoiding "ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all()"